+++ packages/nimble-components/src/table/template.ts
@@ -1,9 +1,29 @@
-import { html, repeat, ref } from '@microsoft/fast-element';
+import {
+    ElementsFilter,
+    html,
+    ref,
+    repeat,
+    slotted,
+    when
+} from '@microsoft/fast-element';
 import { DesignSystem } from '@microsoft/fast-foundation';
 import type { VirtualItem } from '@tanstack/virtual-core';
 import type { Table } from '.';
+import type { TableRowState } from './types';
 import { TableHeader } from './components/header';
 import { TableRow } from './components/row';
+import { TableColumn } from '../table-column/base';
+
+const isTableColumn = (): ElementsFilter => {
+    const filter: ElementsFilter = (
+        value: Node,
+        _: number,
+        __: Node[]
+    ): boolean => {
+        return value instanceof TableColumn;
+    };
+    return filter;
+};
 
 // prettier-ignore
 export const template = html<Table>`
@@ -11,26 +31,29 @@ export const template = html<Table>`
         <div class="table-container">
             <div role="rowgroup" class="header-container" ${ref('headerContainer')} style="margin-right: ${x => x.headerContainerMarginRight}px;">
                 <div class="header-row" role="row">
-                    ${repeat(x => x.columnHeaders, html<string>`
+                    ${repeat(x => x.columns, html<TableColumn>`
                         <${DesignSystem.tagFor(TableHeader)} class="header">
-                            ${x => x}
+                            ${x => x.textContent}
                         </${DesignSystem.tagFor(TableHeader)}>
                     `)}
                 </div>
             </div>
-            <div class="table-viewport" role="rowgroup" ${ref('viewport')}>
+            <div class="table-viewport" ${ref('viewport')}>
+            ${when(x => x.columns.length > 0, html<Table>`
                 <div class="table-scroll" style="height: ${x => x.rowContainerHeight}px;"></div>
-                <div class="table-row-container" ${ref('rowContainer')}>
-                    ${repeat(x => x.visibleItems, html<VirtualItem>`
+                <div class="table-row-container" role="rowgroup" ${ref('rowContainer')}>
+                    ${repeat(x => x.tableData, html<TableRowState>`
                         <${DesignSystem.tagFor(TableRow)}
-                            :data="${(x, c) => (c.parent as Table).data[x.index]}"
+                            :dataRecord="${x => x.record}"
                             :columns="${(_, c) => (c.parent as Table).columns}"
                             style="height: ${x => x.size}px;"
                         >
                         </${DesignSystem.tagFor(TableRow)}>
                     `)}
                 </div>
+            `)}
             </div>
         </div>
+        <slot ${slotted({ property: 'columns', filter: isTableColumn() })}></slot>
     </template>
 `;